cephadm: move iscsi class to a new file
