+++ src/cephadm/cephadmlib/daemons/__init__.py
@@ -2,8 +2,10 @@ from .custom import CustomContainer
 from .tracing import Tracing
 from .ingress import HAproxy, Keepalived
 from .nvmeof import CephNvmeof
+from .iscsi import CephIscsi
 
 __all__ = [
+    'CephIscsi',
     'CephNvmeof',
     'CustomContainer',
     'HAproxy',
